video player and responsive
